feat(client): add mangaers to class
